Feat: Allows the extractor operator's prompt to reference the output of an upstream operator #9869
